refactor: move ParaxialRayTracer to a separate module and update Paraxial class to use it
