+++ optiland/paraxial.py
@@ -8,7 +8,7 @@ Kramer Harrison, 2024
 
 import numpy as np
 
-from optiland.rays import ParaxialRays
+from optiland.raytrace.paraxial_ray_tracer import ParaxialRayTracer
 
 
 class Paraxial:
@@ -30,6 +30,7 @@ class Paraxial:
 
     def __init__(self, optic):
         self.optic = optic
+        self._ray_tracer = ParaxialRayTracer(self.optic)
 
     @property
     def surfaces(self):
@@ -324,53 +325,6 @@ class Paraxial:
 
         return self._trace_generic(-yn[-1], un[-1], z0, wavelength)
 
-    def _get_object_position(self, Hy, y1, EPL):
-        """Calculate the position of the object in the paraxial optical system.
-
-        Args:
-            Hy (float): The normalized field height.
-            y1 (ndarray): The initial y-coordinate of the ray.
-            EPL (float): The effective focal length of the lens.
-
-        Returns:
-            tuple: A tuple containing the y and z coordinates of the object
-                position.
-
-        Raises:
-            ValueError: If the field type is "object_height" and the object is
-                at infinity.
-
-        """
-        obj = self.optic.object_surface
-        field_y = self.optic.fields.max_field * Hy
-
-        if obj.is_infinite:
-            if self.optic.field_type == "object_height":
-                raise ValueError(
-                    'Field type cannot be "object_height" for an object at infinity.',
-                )
-
-            y = -np.tan(np.radians(field_y)) * EPL
-            z = self.optic.surface_group.positions[1]
-
-            y0 = y1 + y
-            z0 = np.ones_like(y1) * z
-        elif self.optic.field_type == "object_height":
-            y = -field_y
-            z = obj.geometry.cs.z
-
-            y0 = np.ones_like(y1) * y
-            z0 = np.ones_like(y1) * z
-
-        elif self.optic.field_type == "angle":
-            y = -np.tan(np.radians(field_y))
-            z = self.optic.surface_group.positions[0]
-
-            y0 = y1 + y
-            z0 = np.ones_like(y1) * z
-
-        return y0, z0
-
     def trace(self, Hy, Py, wavelength):
         """Trace paraxial ray through the optical system based on specified field
         and pupil coordinates.
@@ -381,16 +335,7 @@ class Paraxial:
             wavelength (float): Wavelength of the light.
 
         """
-        EPL = self.EPL()
-        EPD = self.EPD()
-
-        y1 = Py * EPD / 2
-
-        y0, z0 = self._get_object_position(Hy, y1, EPL)
-        u0 = (y1 - y0) / (EPL - z0)
-        rays = ParaxialRays(y0, u0, z0, wavelength)
-
-        self.optic.surface_group.trace(rays)
+        return self._ray_tracer.trace(Hy, Py, wavelength)
 
     def _trace_generic(self, y, u, z, wavelength, reverse=False, skip=0):
         """Trace generically-defined paraxial rays through the optical system.
@@ -410,27 +355,4 @@ class Paraxial:
                 rays after tracing.
 
         """
-        self._process_input(y)
-        self._process_input(u)
-        self._process_input(z)
-
-        surfaces = self.surfaces.inverted() if reverse else self.surfaces
-
-        rays = ParaxialRays(y, u, z, wavelength)
-        surfaces.trace(rays, skip)
-
-        return surfaces.y, surfaces.u
-
-    def _process_input(self, x):
-        """Process input to ensure it is a numpy array.
-
-        Args:
-            x (float or array-like): The input to process.
-
-        Returns:
-            np.ndarray: The processed input.
-
-        """
-        if np.isscalar(x):
-            return np.array([x], dtype=float)
-        return np.array(x, dtype=float)
+        self._ray_tracer.trace_generic(y, u, z, wavelength, reverse, skip)
